Took my edits out
This file has the two sets of downloads currently available from Second Life. They're not complete, and the Bento armatures are, as the website puts it "broken".
I had started editing them to make them workable, and decided to put experimenting into opensim_armaturesBlender2.8xWIP instead. Keep this file as an archive.