Added another SL avatar
Now we have the three 2016 Bento avatars (male, female, angel), and two 2014 "fitted" avatars - armatures and mesh - in this file, as close to the original SL downloads as is possible with Blender 2.8x importing and appending. I relabeled so we can tell the objects apart when we use this file as a source for appending to newer Blender versions.  Licensing and URLs are in a text file within.