removing - need to change direction
the armature is working fine with Appearance sliders inworld, but is distorting the mesh on import if there are any verts weighted to collision bones. Not obvious in the test models - the worst of the bad deform is solved. But it's there, obvious in Roth.  Part of the problem is +X forward vs. +Y forward orientation in Blender before exporting, not completely solved by the Blender-provided SL-OS rigged mesh export settings.  Part of it is in bone roll settings. The rest probably in the bind_mat and rest_mat settings imported with the bento dae files. I'll try going back to my original rebuilt armature and add in cust props with more tests as I go, til I find the bug.