clean up the text file
I changed my mind (and might again) on whether it's better to put the armatures +Y forward in Blender with -X setting on collada export, or +X forward in Blender with +Y settings on collada export. Probably the former (because Blender only mirrors armature settings on the X axis), which means I need to check the X, Y coordinates again in the custom property settings.