Corrections to the rebuilt armature
Instead of just stashing this as a reference, I started using it as a tool. The bone positions fit the default system male a little better than other armatures, for automatic bone weighting, and the bone layers and groups (based on avatar_skeleton.xml 's organization) are easier to use for weight painting than Avastar's armature. So I corrected some of the bone settings to make it work better.