added an armature
v3 is the same as avatar_skeleton.xml with a few changes to make it easier to use for weighting in Blender. I'm using V3 for skinning and weighting, then rotating the mesh to +X forward (apply rotation) and reskinning to avatar_skeleton for exporting.
The changes:
    orient to +Y forward
    add non-deforming bone - PelvisChild - so hips rotate correctly
    change bone names so symmetry works with Blender's naming conventions (as of 2.91)
    clear bone rolls to 0

Problems: I'm still testing this - more nondeform bones may be needed to get the pelvis working correctly. I'm having problems with Blender symmetry on the bones that are in right-left order in the bone hierarchy, and I haven't figured out how to move them nondestructively.