replaced viewer default avatar UV maps
symmetrized, better labels.
Note that the actual viewer default avatars are not symmetrical, though the differences are small. But visible on the UV maps, which is annoying to work with. So I symmetrized a set of set of models in Blender from LL and Cybernetic sources, and exported a new set.